requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-filter==25.1
openpyxl==3.1.5
pillow==11.2.1
pandas==2.2.2
sqlparse==0.5.3
djangorestframework==3.15.1
